Share deterministic waveform generation across metric tests
