+++ test/__init__.py
@@ -1,0 +1,1 @@
+"""VERSA tests and shared test utilities."""
